完成验证码，token认证

+++ mango-admin/src/main/java/com/example/demo/model/Config.java
@@ -99,5 +99,5 @@ public class Config implements Serializable {
      */
     @TableLogic
     @TableField("del_flag")
-    private Byte delFlag;
+    private Integer delFlag;
 }
