+++ mango-admin/src/main/java/com/example/demo/model/Menu.java
@@ -105,5 +105,5 @@ public class Menu implements Serializable {
      */
     @TableLogic
     @TableField("del_flag")
-    private Byte delFlag;
+    private Integer delFlag;
 }
