+++ mango-admin/src/main/java/com/example/demo/model/Role.java
@@ -75,5 +75,5 @@ public class Role implements Serializable {
      */
     @TableLogic
     @TableField("del_flag")
-    private Byte delFlag;
+    private Integer delFlag;
 }
